requirements.txt
flask
numpy==2.0.2
scikit-learn
scipy 
matplotlib
torch
open-clip-torch
pandas